requirements.txt
fastapi
uvicorn
python-dotenv
aiohttp
requests
python-multipart
jinja2